+++ apps/admin-portal/src/app/(home)/page.tsx
@@ -1,3 +1,5 @@
+import { UserType } from "@prisma/client";
+
 import { api } from "../../trpc/server";
 import { CorporateAnalyticsDashboard } from "./_components/corporate-analytics-dashboard";
 import { EmployeeHome } from "./_components/employee-home";
@@ -8,12 +10,12 @@ export default async function Page() {
     const userTypeResult = await api.user.getUserType();
 
     // If user is an employee, show employee home page
-    if (userTypeResult === "EMPLOYEE") {
+    if (userTypeResult === UserType.EMPLOYEE) {
       return <EmployeeHome />;
     }
 
     // If user is corporate, show analytics dashboard
-    if (userTypeResult === "CORPORATE") {
+    if (userTypeResult === UserType.CORPORATE) {
       return <CorporateAnalyticsDashboard />;
     }
