+++ apps/admin-portal/src/app/api/uploadthing/core.ts
@@ -17,8 +17,6 @@ export const ourFileRouter = {
       return { userId: "user-id" }; // Return any metadata you want to pass to onUploadComplete
     })
     .onUploadComplete(async ({ metadata, file }) => {
-      console.log("Upload complete for userId:", metadata.userId);
-      console.log("file url", file.url);
       return { uploadedBy: metadata.userId };
     }),
 } satisfies FileRouter;
