+++ apps/admin-portal/src/middleware.ts
@@ -1,4 +1,3 @@
-import { NextResponse } from "next/server";
 import { clerkMiddleware, createRouteMatcher } from "@clerk/nextjs/server";
 
 const isPublicRoute = createRouteMatcher(["/sign-in(.*)", "/api(.*)"]);
